+++ app/Http/Controllers/TurnoController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Client;
 use App\Models\Marca;
 use App\Models\ModelType;
 use App\Models\StartTurnTime;
@@ -27,48 +28,44 @@ class TurnoController extends Controller
 
     public function index(Request $request)
     {
-        if (is_null(auth()->user()->systemuser)) {
-            return redirect('/usuario');
-        }
+        $this->authorize('viewAny', Turno::class);
+
         $fecha = request()->get('fecha');
-        if($fecha){
-            $turnos = Turno::where('fecha',$fecha)->paginate(5);
-        }else{
+        if ($fecha) {
+            $turnos = Turno::where('fecha', $fecha)->paginate(5);
+        } else {
             $turnos = Turno::paginate(5);
         }
-        return view('turnos.index',compact('turnos'));
+        return view('turnos.index', compact('turnos'));
     }
 
-    public function create(User $user)
+    public function create()
     {
-        $this->authorize('create');
+        $this->authorize('create', Turno::class);
 
         $horarios = StartTurnTime::all();
         $marcas = Marca::all();
         $modelos = ModelType::all();
         $tipos_autos = VehicleType::all();
         $tipo_lavados = WashType::all();
-        return view('turnos.create', compact('horarios', 'marcas', 'modelos', 'tipos_autos', 'tipo_lavados', 'user'));
+        $client_id = request('client_id');
+        return view('turnos.create', compact('horarios', 'marcas', 'modelos', 'tipos_autos', 'tipo_lavados', 'client_id'));
     }
 
-    public function store(User $user)
+    public function store()
     {
+        $this->authorize('create', Turno::class);
+
         if (count(Turno::withTrashed()->latest()->get()) > 0) {
             $identificador = 'T' . sprintf("%'.06d", intval(substr(Turno::withTrashed()->latest()->get()->first()->identificador, 1)) + 1);
         } else {
             $identificador = 'T' . sprintf("%'.06d", 1);
         }
 
-        if (auth()->user()->client) {
-            $client_id = auth()->user()->id;
-        } else
-            $client_id = $user->id;
-
         $turno = Validator::make(array_merge(
             request()->all(),
             [
                 'identificador' => $identificador,
-                'client_id' => $client_id,
                 'precio' => DB::table('wash_types')->where('tipo_lavado', request()['lavado'])->value('precio'),
             ]
         ), [
@@ -84,15 +81,18 @@ class TurnoController extends Controller
             'identificador' => ['unique:turnos'],
             'client_id' => ['required'],
         ])->validate();
-        $user = User::findOrFail($client_id);
-        $user->client->turnos()->create($turno);
 
-        return redirect('/cliente/'.$client_id);
+        $client_id = request('client_id');
+
+        $client = Client::findOrFail($client_id);
+        $client->turnos()->create($turno);
+
+        return redirect()->route('client.show', ['client' => $client_id]);
     }
 
     public function edit(Turno $turno)
     {
-        // $this->authorize('update', $user);
+        $this->authorize('update', $turno);
         $horarios = StartTurnTime::all();
         $marcas = Marca::all();
         $modelos = ModelType::all();
@@ -103,6 +103,7 @@ class TurnoController extends Controller
 
     public function update(Turno $turno)
     {
+        $this->authorize('update', $turno);
 
         $update_turno = Validator::make(array_merge(
             request()->all(),
@@ -124,13 +125,14 @@ class TurnoController extends Controller
 
         $turno->update($update_turno);
 
-        return redirect('/cliente/' . $turno->client->user->id);
+        return redirect()->route('client.show', ['client' => $turno->client_id]);
     }
 
-    public function delete(Turno $turno)
+    public function destroy(Turno $turno)
     {
-        $client_id = $turno->client->user_id;
+        $this->authorize('delete', $turno);
+        $client_id = $turno->client_id;
         $turno->delete();
-        return redirect('/cliente/' . $client_id);
+        return redirect()->route('client.show', ['client' => $turno->client_id]);
     }
 }
